Warning fixes.

+++ CIAM/LoggerFactory.cpp
@@ -46,7 +46,7 @@ void LoggerFactory::XMLParse( const DOMNode* root ) {
 	nodeList = root->getChildNodes();
 	
 	// loop through the children
-	for ( int i = 0; i < nodeList->getLength(); i++ ){
+	for ( int i = 0; i < static_cast<int>( nodeList->getLength() ); i++ ){
 		curr = nodeList->item( i );
 		nodeName = XMLHelper<string>::safeTranscode( curr->getNodeName() );
