+++ CIAM/modeltime.cpp
@@ -80,7 +80,7 @@ void Modeltime::XMLParse( const DOMNode* node ) {
 	nodeList = node->getChildNodes();
 	
 	// loop through the children
-	for ( int i = 0; i < nodeList->getLength(); i++ ){
+	for ( int i = 0; i < static_cast<int>( nodeList->getLength() ); i++ ){
 		curr = nodeList->item( i );
 		nodeName = XMLHelper<string>::safeTranscode( curr->getNodeName() );
